Moved managed_error into error pkg
